Created separate app for reset pusword functionality with Servise class
