+++ backend/socialmedia/accounts/serializers.py
@@ -74,77 +74,6 @@ class LoginSerializer(serializers.ModelSerializer):
         }
 
 
-class PasswordResetRequestSerializer(serializers.Serializer):
-    email = serializers.EmailField(max_length=255, min_length=6)
-
-    class Meta:
-        fields = ["email"]
-
-    def validate(self, attrs):
-        email = attrs.get("email", "")
-
-        if User.objects.filter(email=email).exists():
-            user = User.objects.get(email=email)
-            uidb64 = urlsafe_base64_encode(smart_bytes(user.id))
-            token = PasswordResetTokenGenerator().make_token(user)
-            request = self.context.get("request")
-
-            side_domain = get_current_site(request).domain
-            relative_link = reverse(
-                "password-reset-confirm", kwargs={"uidb64": uidb64, "token": token}
-            )
-
-            abslink = f"http://{side_domain}{relative_link}"
-            email_body = (
-                f"Hello, \n Use the link below to reset your password \n {abslink}"
-            )
-            data = {
-                "email_body": email_body,
-                "email_subject": "Reset your password",
-                "to_email": user.email,
-            }
-
-            send_normal_email(data)
-
-        return attrs
-
-
-class SetNewPasswordSerializer(serializers.Serializer):
-    password = serializers.CharField(max_length=68, min_length=6, write_only=True)
-    confirm_password = serializers.CharField(
-        max_length=68, min_length=6, write_only=True
-    )
-    token = serializers.CharField(write_only=True)
-    uidb64 = serializers.CharField(write_only=True)
-
-    class Meta:
-        fields = ["password", "confirm_password", "token", "uidb64"]
-
-    def validate(self, attrs):
-        try:
-            password = attrs.get("password", "")
-            confirm_password = attrs.get("confirm_password", "")
-            token = attrs.get("token", "")
-            uidb64 = attrs.get("uidb64", "")
-
-            user_id = force_str(urlsafe_base64_decode(uidb64))
-            user = User.objects.get(id=user_id)
-
-            if not PasswordResetTokenGenerator().check_token(user, token):
-                raise AuthenticationFailed("The reset link is invalid", 401)
-
-            if password != confirm_password:
-                raise AuthenticationFailed("Passwords do not match", 401)
-
-            user.set_password(password)
-            user.save()
-
-            return user
-
-        except Exception as e:
-            raise AuthenticationFailed("The reset link is invalid", 401)
-
-
 class LogoutUserSerializer(serializers.Serializer):
     refresh_token = serializers.CharField()
