+++ backend/socialmedia/password_reset/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class PasswordResetConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'password_reset'
