+++ backend/socialmedia/socialmedia/urls.py
@@ -4,4 +4,5 @@ from django.urls import path, include
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("api/v1/auth/", include("accounts.urls")),
+    path("api/v1/auth/", include("password_reset.urls")),
 ]
